* Added addition classNames to be attached to BaseModal. * Refactored example files * Tested initial large-modal
