+++ frontend/src/components/BaseModal.tsx
@@ -5,13 +5,14 @@ import { RefObject } from "react";
 export interface BaseModalProps {
   children: React.ReactNode;
   modalRef: RefObject<HTMLDialogElement>;
+  classNames?: string;
 }
 
 function BaseModal(props: BaseModalProps) {
-  const { children, modalRef } = props;
+  const { children, modalRef, classNames } = props;
 
   return (
-    <dialog ref={modalRef} className="rounded-lg p-4">
+    <dialog ref={modalRef} className={`rounded-lg p-4 ${classNames}`}>
       {children}
     </dialog>
   );
