+++ frontend/src/components/WeekSelection.tsx
@@ -5,6 +5,8 @@ import { useSelectedWeek } from "@/hooks/staffing/useSelectedWeek";
 import DropDown from "./DropDown";
 import { useUrlRouteFilter } from "@/hooks/staffing/useUrlRouteFilter";
 import ActionButton from "./Buttons/ActionButton";
+import { LargeModal } from "./EasyModal/LargeModal";
+import ButtonExampleModal from "./EasyModal/ButtonExampleModal";
 
 export default function WeekSelection() {
   const weekSpanOptions = ["8 uker", "12 uker", "26 uker"];
@@ -20,6 +22,9 @@ export default function WeekSelection() {
 
   return (
     <div className="flex flex-row gap-2">
+      <LargeModal />
+      <ButtonExampleModal />
+
       <DropDown
         startingOption={weekSpan ? weekSpan + " uker" : weekSpanOptions[0]}
         dropDownOptions={weekSpanOptions}
